feat(frontend): support static export for single-process serving
The app is a client-rendered SPA that talks to the backend same-origin, so
it can ship as static assets served by FastAPI (no Node at runtime) for the
pip/uvx deployment.

- next.config: output mode is env-driven (NEXT_OUTPUT=standalone keeps the
  Node server for the Docker image; default is static 'export')
- Dockerfile.frontend sets NEXT_OUTPUT=standalone to preserve current behavior
- layout: drop request-time cookies()/headers(); I18nProvider already restores
  the locale from localStorage on the client
- dynamic routes ([id]/[token]): split into a server wrapper that supplies a
  placeholder param via generateStaticParams plus the existing client page, so
  export emits a shell HTML that FastAPI serves for any real id
- widget/chat: wrap useSearchParams in a Suspense boundary (required by export)
