+++ frontend/next.config.mjs
@@ -3,10 +3,17 @@ import { fileURLToPath } from "node:url";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
+// Output mode:
+//   'export'     -> static assets in out/, served by FastAPI in the single-process
+//                   (pip / uvx) deployment. This is the default.
+//   'standalone' -> self-contained Node server, used by the multi-container Docker
+//                   deployment (set NEXT_OUTPUT=standalone in Dockerfile.frontend).
+const outputMode = process.env.NEXT_OUTPUT === 'standalone' ? 'standalone' : 'export';
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  // Enable standalone output for Docker deployment
-  output: 'standalone',
+  output: outputMode,
+  images: { unoptimized: true },
   outputFileTracingRoot: __dirname,
   experimental: {
     optimizeCss: false,
