+++ frontend/src/app/share/[token]/page.tsx
@@ -1,16 +1,13 @@
-"use client"
+import React from "react"
+import PageClient from "./page-client"
 
-import { useParams } from "next/navigation"
-import { PublicAgentChatPage } from "@/components/widget/public-agent-chat-page"
+// Server wrapper for static export: provides a placeholder param so a shell
+// HTML is emitted for this dynamic route. FastAPI serves the shell for any real
+// token; the client component reads the actual value from the URL via useParams().
+export function generateStaticParams() {
+  return [{ token: "__shell__" }]
+}
 
-export default function ShareChatPage() {
-  const params = useParams()
-  const token = params.token as string
-
-  return (
-    <PublicAgentChatPage
-      authMode="share"
-      routeToken={token}
-    />
-  )
+export default function Page() {
+  return <PageClient />
 }
